move example into its own file, seperate out functions to use
